fixed the set volume functions in VBS and actuator which was trimming floating point to 0. Also set the volume at the end points so the piston can recalibrate at the end points

+++ firmware/lib/vbs/actuator.cpp
@@ -95,7 +95,7 @@ float Actuator::calculate_motor_power(float motorTorque,  float rotVel)
     return powerConsumption;
 }
 
-void Actuator::set_volume(int vol_mL)
+void Actuator::set_volume(float vol_mL)
 {
     double vol_m3 = vol_mL/1000000;
     _pistonVolume = vol_m3;
